+++ exporter/collector/integrationtest/metric_test_server.go
@@ -16,153 +16,21 @@ package integrationtest
 
 import (
 	"context"
-	"net"
-	"strings"
-	"sync"
 	"testing"
 
 	"github.com/stretchr/testify/require"
 	"go.uber.org/zap"
-	"google.golang.org/genproto/googleapis/api/metric"
-	monitoringpb "google.golang.org/genproto/googleapis/monitoring/v3"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
-	"google.golang.org/protobuf/types/known/emptypb"
 
 	"github.com/GoogleCloudPlatform/opentelemetry-operations-go/exporter/collector"
+	"github.com/GoogleCloudPlatform/opentelemetry-operations-go/internal/cloudmock"
 )
 
-type MetricsTestServer struct {
-	lis                         net.Listener
-	srv                         *grpc.Server
-	Endpoint                    string
-	createMetricDescriptorReqs  []*monitoringpb.CreateMetricDescriptorRequest
-	createTimeSeriesReqs        []*monitoringpb.CreateTimeSeriesRequest
-	createServiceTimeSeriesReqs []*monitoringpb.CreateTimeSeriesRequest
-	mu                          sync.Mutex
-}
-
-func (m *MetricsTestServer) Shutdown() {
-	// this will close mts.lis
-	m.srv.GracefulStop()
-}
-
-// Pops out the CreateMetricDescriptorRequests which the test server has received so far
-func (m *MetricsTestServer) CreateMetricDescriptorRequests() []*monitoringpb.CreateMetricDescriptorRequest {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	reqs := m.createMetricDescriptorReqs
-	m.createMetricDescriptorReqs = nil
-	return reqs
-}
-
-// Pops out the CreateTimeSeriesRequests which the test server has received so far
-func (m *MetricsTestServer) CreateTimeSeriesRequests() []*monitoringpb.CreateTimeSeriesRequest {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	reqs := m.createTimeSeriesReqs
-	m.createTimeSeriesReqs = nil
-	return reqs
-}
-
-// Pops out the CreateServiceTimeSeriesRequests which the test server has received so far
-func (m *MetricsTestServer) CreateServiceTimeSeriesRequests() []*monitoringpb.CreateTimeSeriesRequest {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	reqs := m.createServiceTimeSeriesReqs
-	m.createServiceTimeSeriesReqs = nil
-	return reqs
-}
-
-func (m *MetricsTestServer) appendCreateMetricDescriptorReq(req *monitoringpb.CreateMetricDescriptorRequest) {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	m.createMetricDescriptorReqs = append(m.createMetricDescriptorReqs, req)
-}
-func (m *MetricsTestServer) appendCreateTimeSeriesReq(req *monitoringpb.CreateTimeSeriesRequest) {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	m.createTimeSeriesReqs = append(m.createTimeSeriesReqs, req)
-}
-func (m *MetricsTestServer) appendCreateServiceTimeSeriesReq(req *monitoringpb.CreateTimeSeriesRequest) {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	m.createServiceTimeSeriesReqs = append(m.createServiceTimeSeriesReqs, req)
-}
-
-func (m *MetricsTestServer) Serve() error {
-	return m.srv.Serve(m.lis)
-}
-
-type fakeMetricServiceServer struct {
-	monitoringpb.UnimplementedMetricServiceServer
-	metricsTestServer *MetricsTestServer
-}
-
-func (f *fakeMetricServiceServer) CreateTimeSeries(
-	ctx context.Context,
-	req *monitoringpb.CreateTimeSeriesRequest,
-) (*emptypb.Empty, error) {
-	f.metricsTestServer.appendCreateTimeSeriesReq(req)
-
-	var code codes.Code
-	if strings.Contains(req.Name, "notfound") {
-		code = codes.NotFound
-	}
-
-	statusResp := status.New(code, "FORCED_TEST_ERROR")
-	if code == codes.NotFound {
-		statusResp, _ = statusResp.WithDetails(&monitoringpb.CreateTimeSeriesSummary{
-			TotalPointCount:   int32(len(req.TimeSeries)),
-			SuccessPointCount: 0,
-		})
-	}
-
-	return &emptypb.Empty{}, statusResp.Err()
-}
-
-func (f *fakeMetricServiceServer) CreateServiceTimeSeries(
-	ctx context.Context,
-	req *monitoringpb.CreateTimeSeriesRequest,
-) (*emptypb.Empty, error) {
-	f.metricsTestServer.appendCreateServiceTimeSeriesReq(req)
-	return &emptypb.Empty{}, nil
-}
-
-func (f *fakeMetricServiceServer) CreateMetricDescriptor(
-	ctx context.Context,
-	req *monitoringpb.CreateMetricDescriptorRequest,
-) (*metric.MetricDescriptor, error) {
-	f.metricsTestServer.appendCreateMetricDescriptorReq(req)
-	return &metric.MetricDescriptor{}, nil
-}
-
-func NewMetricTestServer() (*MetricsTestServer, error) {
-	srv := grpc.NewServer()
-	lis, err := net.Listen("tcp", "localhost:0")
-	if err != nil {
-		return nil, err
-	}
-	testServer := &MetricsTestServer{
-		Endpoint: lis.Addr().String(),
-		lis:      lis,
-		srv:      srv,
-	}
-
-	monitoringpb.RegisterMetricServiceServer(
-		srv,
-		&fakeMetricServiceServer{metricsTestServer: testServer},
-	)
-
-	return testServer, nil
-}
-
-// NewExporter creates and starts a googlecloud exporter by updating the
+// NewMetricsTestExporter creates and starts a googlecloud exporter by updating the
 // given cfg copy to point to the test server.
-func (m *MetricsTestServer) NewExporter(
+func NewMetricTestExporter(
 	ctx context.Context,
 	t testing.TB,
+	m *cloudmock.MetricsTestServer,
 	cfg collector.Config,
 ) *collector.MetricsExporter {
 	cfg.MetricConfig.ClientConfig.Endpoint = m.Endpoint
