+++ exporter/collector/integrationtest/trace_test_server.go
@@ -16,84 +16,18 @@ package integrationtest
 
 import (
 	"context"
-	"net"
-	"sync"
 	"testing"
 
 	"github.com/stretchr/testify/require"
-	tracepb "google.golang.org/genproto/googleapis/devtools/cloudtrace/v2"
-	"google.golang.org/grpc"
-	"google.golang.org/protobuf/types/known/emptypb"
 
 	"github.com/GoogleCloudPlatform/opentelemetry-operations-go/exporter/collector"
+	"github.com/GoogleCloudPlatform/opentelemetry-operations-go/internal/cloudmock"
 )
 
-type TracesTestServer struct {
-	lis net.Listener
-	srv *grpc.Server
-	// Endpoint where the gRPC server is listening
-	Endpoint                string
-	batchWriteSpansRequests []*tracepb.BatchWriteSpansRequest
-	mu                      sync.Mutex
-}
-
-func (t *TracesTestServer) Shutdown() {
-	t.srv.GracefulStop()
-}
-
-func (t *TracesTestServer) Serve() {
-	t.srv.Serve(t.lis)
-}
-
-type fakeTraceServiceServer struct {
-	tracepb.UnimplementedTraceServiceServer
-	tracesTestServer *TracesTestServer
-}
-
-func (f *fakeTraceServiceServer) BatchWriteSpans(
-	ctx context.Context,
-	request *tracepb.BatchWriteSpansRequest,
-) (*emptypb.Empty, error) {
-	f.tracesTestServer.appendBatchWriteSpansRequest(request)
-	return &emptypb.Empty{}, nil
-}
-
-func (t *TracesTestServer) appendBatchWriteSpansRequest(req *tracepb.BatchWriteSpansRequest) {
-	t.mu.Lock()
-	defer t.mu.Unlock()
-	t.batchWriteSpansRequests = append(t.batchWriteSpansRequests, req)
-}
-
-func (t *TracesTestServer) CreateBatchWriteSpansRequests() []*tracepb.BatchWriteSpansRequest {
-	t.mu.Lock()
-	defer t.mu.Unlock()
-	reqs := t.batchWriteSpansRequests
-	t.batchWriteSpansRequests = nil
-	return reqs
-}
-
-func NewTracesTestServer() (*TracesTestServer, error) {
-	srv := grpc.NewServer()
-	lis, err := net.Listen("tcp", "localhost:0")
-	if err != nil {
-		return nil, err
-	}
-	testServer := &TracesTestServer{
-		Endpoint: lis.Addr().String(),
-		lis:      lis,
-		srv:      srv,
-	}
-	tracepb.RegisterTraceServiceServer(
-		srv,
-		&fakeTraceServiceServer{tracesTestServer: testServer},
-	)
-
-	return testServer, nil
-}
-
-func (s *TracesTestServer) NewExporter(
+func NewTraceTestExporter(
 	ctx context.Context,
 	t testing.TB,
+	s *cloudmock.TracesTestServer,
 	cfg collector.Config,
 ) *collector.TraceExporter {
