+++ internal/cloudmock/go.mod
@@ -1,0 +1,16 @@
+module github.com/GoogleCloudPlatform/opentelemetry-operations-go/internal/cloudmock
+
+go 1.17
+
+require (
+	google.golang.org/genproto v0.0.0-20220829175752-36a9c930ecbf
+	google.golang.org/grpc v1.49.0
+	google.golang.org/protobuf v1.28.1
+)
+
+require (
+	github.com/golang/protobuf v1.5.2 // indirect
+	golang.org/x/net v0.0.0-20210405180319-a5a99cb37ef4 // indirect
+	golang.org/x/sys v0.0.0-20210510120138-977fb7262007 // indirect
+	golang.org/x/text v0.3.5 // indirect
+)
